src/requirements.txt
PyYAML==6.0.1
requests==2.28.1
urllib3==1.26.10